+++ common/src/main/java/fr/loudo/narrativecraft/narrative/npc/NpcPayload.java
@@ -70,7 +70,7 @@ public class NpcPayload extends NarrativeEntryPayload {
         this.sceneId = sceneId;
         this.chapterId = chapterId;
         this.dialogDataJson = dialogDataJson != null ? dialogDataJson : "{}";
-        this.customNbt = customNbt;
+        this.customNbt = customNbt != null ? customNbt : "";
     }
 
     public String getModelType() {