+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Render image in dialog `Clara: [img clara/happy] I'm feeling good right now!`
 - Item requirement tag
 - Locked choice
+- Character head bobbing when talking (it's back!)
 
 ### Fixes
 - Dialog 2D scale not fixed based on gui scale